Add fixture to provide correct Headers
When updating the TXT records the plugin not only checks the returned
status code of the executed API call but also checks the returned headers
for a success message.

As we do not actually communicate with the API during testing the returned
headers were missing the expected success message, leading to failing tests
when checking the perform() and cleanup() methods.

In order to fix these errors, this commit introduces a ew fixture to the
test suite that ensures a proper (i.e. OK) status code as well as proper
headers containing the expected success information during testing